requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
pip-autoremove==0.9.1
psycopg2==2.7.7
pytz==2018.9
